# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VideoGames API",
        "description": "API REST para la gestión de videojuegos, reseñas, compras y usuarios. Proyecto Laravel 12 con MySQL.",
        "contact": {
            "name": "Equipo VideoGames API",
            "email": "soporte@videogamesapi.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Listar todos los videojuegos",
                "operationId": "e333e2e46e6c7bd4c254fe379b7389dd",
                "responses": {
                    "200": {
                        "description": "Lista de videojuegos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Game"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Crear un nuevo videojuego",
                "operationId": "496a1db2ab2b344c40cad9e98d332d9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Game"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Videojuego creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Obtener detalles de un videojuego",
                "operationId": "48fb06317bf9e8f10c22dd892ff7425e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del videojuego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del videojuego",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Videojuego no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Games"
                ],
                "summary": "Actualizar un videojuego",
                "operationId": "273f111abbd78231f5f9c48d4fca792c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del videojuego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Game"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Videojuego actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Videojuego no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Eliminar un videojuego",
                "operationId": "c5bffb910bb1804bac14c04327dad425",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del videojuego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado correctamente"
                    },
                    "404": {
                        "description": "Videojuego no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Listar todos los géneros",
                "operationId": "0ff78353b1c27ef8905e58f3cf31370a",
                "responses": {
                    "200": {
                        "description": "Lista de géneros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Genre"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Crear un nuevo género",
                "operationId": "8d21057e41193b21373f4e2cce0a8c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Genre"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Género creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genres/{id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Obtener un género",
                "operationId": "cec9d00fae56cff69cba705f6b436983",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Genres"
                ],
                "summary": "Actualizar un género",
                "operationId": "9ec4944f7fd50bd7280fb1366c748709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Genre"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Eliminar un género",
                "operationId": "1890f1018746d6e4385aa0c3244acc08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Genre": {
                "title": "Genre",
                "description": "Género de videojuego",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acción"
                    },
                    "description": {
                        "type": "string",
                        "example": "Juegos con enfoque en el combate."
                    }
                },
                "type": "object"
            },
...
            "Game": {
                "title": "Game",
                "description": "Videojuego",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Last of Us Part II"
                    },
                    "description": {
                        "type": "string",
                        "example": "Juego de acción y aventura post-apocalíptico."
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2020-06-19"
                    },
                    "genre_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "platform_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "developer_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "publisher_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}